Move tests to ReactPropTypes-test.js
